ai-python/requirements.txt
# ===== Web & runtime =====
fastapi==0.115.5
uvicorn[standard]==0.32.1
httpx==0.27.2
pydantic==2.9.2
pydantic-settings==2.6.1
requests==2.32.3
pymongo==4.10.1
python-multipart==0.0.9
hf-transfer==0.1.6           # <-- NEW: provides `hf_transfer` module for fast downloads

# ===== HF / Transformers / Embeddings / RAG helpers =====
transformers==4.44.2
accelerate==0.33.0
sentence-transformers==3.2.1
einops==0.8.0
sentencepiece==0.2.0
huggingface-hub==0.26.2
safetensors==0.4.5
faiss-cpu==1.8.0.post1
tqdm==4.67.1
tokenizers==0.19.1

# ===== Vision/IO & docs =====
pillow==10.4.0
PyPDF2==3.0.1
python-docx==1.1.2
openpyxl==3.1.5
opencv-python-headless==4.10.0.84
opencc-python-reimplemented==0.1.7

# ===== Low-level compat =====
numpy==1.26.4
